requirements.txt
asgiref==3.7.2
autopep8==2.0.2
blinker==1.6.2
click==8.1.4
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
dnspython==2.3.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycodestyle==2.10.0
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
Werkzeug==2.3.6
